Add new feature, wks.get_all_records()
wks.get_all_records() treats the spreadsheet like a lightweight database,
returning all rows after the first one as a dictionary that has the first
row of values as keys.

In addition, values are not merely returned as strings, but converted when
possible to a float or an int. A boolean param called empty2zero controls
whether empty cells are returned as empty strings or as zeroes.