Improvement: Fetch SpreadSheet creation and lastUpdate time.
On a call to `open` or `openall` methods we use the GoogleDrive API,
set the list of fields to retrieve and retrieve `createdTime` and `modifiedTime`.
This allows us to get the creation and last modified time of a SpreadSheet.

When using `open_by_key` or `open_by_url` we do not use the GoogleDrive API,
this prevent us from accessing this informaiton, if a user tries to access the `creationTime` or
`lastUpdateTime` property of a SpreadSheet then query the GoogleDrive API to get it.

Benefit from that call to update SpreadSheet metadata on the way.

Signed-off-by: Alexandre Lavigne <lavigne958@gmail.com>