Fix client.copy copy_permission argument compare
When comparing value in `if` statement, when that value
should be a boolean use `if value is True :` instead of `if value :`
because any type of variable/object would meet the `True` condition
in the second case but only `True` would meet the `True` condition
on the first case.

In the case of `copy_permission` we expect a boolean of value `True`
only.

closes #892