Simplify cassette recording for test setup
- Use a single cassette for the whole init/teardown
- Use a cassette file name with no white characters

If test needs to update the API call it makes, it is easier to only have
to re-generate a single file.

The spreadsheet creation is done only once but the single sheet is
cleared everytime before a new test starts. We must record the
spreadsheet creation then close the cassette to allow the fixture
`reset_sheet` to find the cassette. Then open it again to record the
file deletion.

Next time a test needs to be updated and make an extra API call,
simplify delete the cassette and run the test recording new episodes.

Closes #981