Fix test cassette management
Before each test, the current worksheet is cleared to make
sure the test start with no values from previous test.

This was recorded in the same test cassette and the API
call to clear the sheet is the same every time. This leads
to pyvcr not making the actuall HTTP request and not clearing
the worksheet because it uses the response from the cassette.

Moved the record of the HTTP requests to clear the worksheet
with the test itself, this makes sure it is made for each test
during recording.

Closes #983