Add exponential backoff client
Add a new client that provides exponential backoff retries.
This is usefull for the test suite to retry the requests
if they encounter an API usage rate limit.

This `Client` instance is not producation ready yet.
To use with caution.

I can make the request retry up to a couple of minutes for requests
that should fail on the first try.
The current list of matching codes when the client should retry
is not exhaustive enough and will be improved in the future.