Add method to hide rows/columns
Add new method that allows the user to explicitly hide rows or columns.
Ass new method that allows the user to explicitly unhide rows or
columns.

This feature does not include any _listing_ method, because the listing
is not in the metadata of the `SpreadSheet` or the `Sheet`, it will
require to list all the data in the sheet and list all of them
then check if the flag `hiddenByUser` is set. This is a lot of
processing and with very large data set this can take some time
For a result that is only visible using the UI.

closes #1006